pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "feishu-campus-longmemory"
version = "1.0.0"
description = "Personal Work Memory Middleware for Feishu and OpenClaw."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "alembic>=1.13,<2.0",
    "fastapi>=0.111,<1.0",
    "lark-oapi>=1.5,<2.0",
    "pydantic-settings>=2.2,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "tzdata>=2024.1",
    "uvicorn[standard]>=0.29,<1.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27,<1.0",
    "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
